feat(throttleTime): add higher-order lettable version of throttleTime
