+++ src/operator/throttleTime.ts
@@ -1,10 +1,8 @@
-import { Operator } from '../Operator';
-import { Subscriber } from '../Subscriber';
-import { IScheduler } from '../Scheduler';
-import { Subscription, TeardownLogic } from '../Subscription';
-import { async } from '../scheduler/async';
 import { Observable } from '../Observable';
-import { ThrottleConfig, defaultThrottleConfig } from './throttle';
+import { IScheduler } from '../Scheduler';
+import { async } from '../scheduler/async';
+import { ThrottleConfig, defaultThrottleConfig } from '../operators/throttle';
+import { throttleTime as higherOrder } from '../operators/throttleTime';
 
 /**
  * Emits a value from the source Observable, then ignores subsequent source
@@ -49,75 +47,5 @@ export function throttleTime<T>(this: Observable<T>,
                                 duration: number,
                                 scheduler: IScheduler = async,
                                 config: ThrottleConfig = defaultThrottleConfig): Observable<T> {
-  return this.lift(new ThrottleTimeOperator(duration, scheduler, config.leading, config.trailing));
-}
-
-class ThrottleTimeOperator<T> implements Operator<T, T> {
-  constructor(private duration: number,
-              private scheduler: IScheduler,
-              private leading: boolean,
-              private trailing: boolean) {
-  }
-
-  call(subscriber: Subscriber<T>, source: any): TeardownLogic {
-    return source.subscribe(
-      new ThrottleTimeSubscriber(subscriber, this.duration, this.scheduler, this.leading, this.trailing)
-    );
-  }
-}
-
-/**
- * We need this JSDoc comment for affecting ESDoc.
- * @ignore
- * @extends {Ignored}
- */
-class ThrottleTimeSubscriber<T> extends Subscriber<T> {
-  private throttled: Subscription;
-  private _hasTrailingValue: boolean = false;
-  private _trailingValue: T = null;
-
-  constructor(destination: Subscriber<T>,
-              private duration: number,
-              private scheduler: IScheduler,
-              private leading: boolean,
-              private trailing: boolean) {
-    super(destination);
-  }
-
-  protected _next(value: T) {
-    if (this.throttled) {
-      if (this.trailing) {
-        this._trailingValue = value;
-        this._hasTrailingValue = true;
-      }
-    } else {
-      this.add(this.throttled = this.scheduler.schedule(dispatchNext, this.duration, { subscriber: this }));
-      if (this.leading) {
-        this.destination.next(value);
-      }
-    }
-  }
-
-  clearThrottle() {
-    const throttled = this.throttled;
-    if (throttled) {
-      if (this.trailing && this._hasTrailingValue) {
-        this.destination.next(this._trailingValue);
-        this._trailingValue = null;
-        this._hasTrailingValue = false;
-      }
-      throttled.unsubscribe();
-      this.remove(throttled);
-      this.throttled = null;
-    }
-  }
-}
-
-interface DispatchArg<T> {
-  subscriber: ThrottleTimeSubscriber<T>;
-}
-
-function dispatchNext<T>(arg: DispatchArg<T>) {
-  const { subscriber } = arg;
-  subscriber.clearThrottle();
+  return higherOrder(duration, scheduler, config)(this);
 }
