+++ src/operators/index.ts
@@ -66,6 +66,7 @@ export { switchMap } from './switchMap';
 export { takeLast } from './takeLast';
 export { tap } from './tap';
 export { throttle } from './throttle';
+export { throttleTime } from './throttleTime';
 export { timestamp } from './timestamp';
 export { toArray } from './toArray';
 export { window } from './window';
